Refactor EditPass view
This change makes this view not a Form and makes the EditablePassCard
view not a header to the form, which was cumbersome and probably not
best practice.

This will require the pickers used in this view to be updated to mimic
the appearance they're given when used inside of a Form.